+++ src/update_time/updaters/update_manifest_images.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from update_time.io.filesystem import YAML_GLOB_PATTERNS, update_files
 from update_time.io.log import get_logger
-from update_time.sources.docker import IMAGE_REFERENCE, get_latest_tag
+from update_time.sources.oci import IMAGE_REFERENCE, get_latest_tag
 
 LOG = get_logger("manifest images")
 IMAGE_RE = rf"image: {IMAGE_REFERENCE}"